Cache architecture in the whole app
