--- lib/datasource/local/cache_keys.dart
@@ -1,6 +1,0 @@
-class CacheKeys {
-  static String proposalVotesCacheName = "votes.2.box";
-  static String voteCacheKey(String account, int proposalId) {
-    return "${account}_$proposalId";
-  }  
-}
