+++ lib/datasource/remote/model/vote_model.dart
@@ -1,5 +1,3 @@
-import 'package:hive/hive.dart';
-
 /// The user's vote
 class VoteModel {
   final int amount;
@@ -14,28 +12,5 @@ class VoteModel {
     } else {
       return const VoteModel(amount: 0, isVoted: false);
     }
-  }
-}
-
-class VoteModelAdapter extends TypeAdapter<VoteModel> {
-  @override
-  final typeId = 3;
-
-  @override
-  VoteModel read(BinaryReader reader) {
-    final fields = [];
-    reader.readByte();
-    fields.add(reader.readInt());
-    reader.readByte();
-    fields.add(reader.readBool());
-    return VoteModel(amount: fields[0], isVoted: fields[1]);
-  }
-
-  @override
-  void write(BinaryWriter writer, VoteModel obj) {
-    writer.writeByte(0);
-    writer.writeInt(obj.amount);
-    writer.writeByte(1);
-    writer.writeBool(obj.isVoted);
   }
 }
