+++ lib/main.dart
@@ -11,10 +11,10 @@ import 'package:seeds/blocs/authentication/viewmodels/bloc.dart';
 import 'package:seeds/blocs/deeplink/viewmodels/deeplink_bloc.dart';
 import 'package:seeds/blocs/deeplink/viewmodels/deeplink_state.dart';
 import 'package:seeds/datasource/local/member_model_cache_item.dart';
+import 'package:seeds/datasource/local/models/vote_model_adapter.dart';
 import 'package:seeds/datasource/local/settings_storage.dart';
 import 'package:seeds/datasource/remote/firebase/firebase_push_notification_service.dart';
 import 'package:seeds/datasource/remote/firebase/firebase_remote_config.dart';
-import 'package:seeds/datasource/remote/model/vote_model.dart';
 import 'package:seeds/domain-shared/bloc_observer.dart';
 import 'package:seeds/navigation/navigation_service.dart';
 import 'package:seeds/screens/app/app.dart';
