+++ lib/screens/explore_screens/vote_screens/proposal_details/interactor/usecases/get_proposal_data_use_case.dart
@@ -1,12 +1,12 @@
 import 'package:async/async.dart';
 import 'package:hive/hive.dart';
+import 'package:seeds/datasource/local/cache_repository.dart';
 import 'package:seeds/datasource/local/settings_storage.dart';
 import 'package:seeds/datasource/remote/api/profile_repository.dart';
 import 'package:seeds/datasource/remote/api/proposals_repository.dart';
 import 'package:seeds/datasource/remote/api/voice_repository.dart';
 import 'package:seeds/datasource/remote/model/proposals_model.dart';
 import 'package:seeds/datasource/remote/model/vote_model.dart';
-import 'package:seeds/datasource/local/cache_keys.dart';
 
 class GetProposalDataUseCase {
   final ProfileRepository _profileRepository = ProfileRepository();
@@ -28,14 +28,15 @@ class GetProposalDataUseCase {
   }
 
   Future<Result> _fetchVote(int proposalId, String account) async {
-    final box = await Hive.openBox<VoteModel>(CacheKeys.proposalVotesCacheName);
-    VoteModel? voteModel = box.get(CacheKeys.voteCacheKey(account, proposalId));
+    final box = await Hive.openBox<VoteModel>(proposalVotesCacheBox);
+    final cache = CacheRepository<VoteModel>(box);
+    VoteModel? voteModel = cache.get(buildVoteKey(account, proposalId));
     if (voteModel == null) {
       final result = await _proposalsRepository.getVote(proposalId, account);
       if (result.isValue) {
         voteModel = result.asValue!.value as VoteModel;
         if (voteModel.isVoted) {
-          await box.put(proposalId, voteModel);
+          await cache.add(proposalId, voteModel);
         }
       }
     }
