+++ lib/screens/explore_screens/vote_screens/proposals/components/vote_amount_label/interactor/usecases/get_vote_use_case.dart
@@ -3,13 +3,12 @@ import 'package:hive/hive.dart';
 import 'package:seeds/datasource/local/cache_repository.dart';
 import 'package:seeds/datasource/remote/api/proposals_repository.dart';
 import 'package:seeds/datasource/remote/model/vote_model.dart';
-import 'package:seeds/datasource/local/cache_keys.dart';
 
 class GetVoteUseCase {
   Future<Result> run(int proposalId, String account) async {
-    final box = await Hive.openBox<VoteModel>(CacheKeys.proposalVotesCacheName);
+    final box = await Hive.openBox<VoteModel>(proposalVotesCacheBox);
     final cache = CacheRepository<VoteModel>(box);
-    VoteModel? voteModel = box.get(CacheKeys.voteCacheKey(account, proposalId));
+    VoteModel? voteModel = cache.get(buildVoteKey(account, proposalId));
     if (voteModel == null) {
       final result = await ProposalsRepository().getVote(proposalId, account);
       if (result.isValue) {
